Use long for hashing the EntryStack

+++ api/src/main/java/me/shedaniel/rei/api/common/entry/type/EntryDefinition.java
@@ -59,7 +59,7 @@ public interface EntryDefinition<T> {
     
     T normalize(EntryStack<T> entry, T value);
     
-    int hash(EntryStack<T> entry, T value, ComparisonContext context);
+    long hash(EntryStack<T> entry, T value, ComparisonContext context);
     
     boolean equals(T o1, T o2, ComparisonContext context);
