+++ default-plugin/src/main/java/me/shedaniel/rei/plugin/client/favorites/WeatherFavoriteEntry.java
@@ -170,7 +170,7 @@ public class WeatherFavoriteEntry extends FavoriteEntry {
     }
     
     @Override
-    public int hashIgnoreAmount() {
+    public long hashIgnoreAmount() {
         return weather == null ? -1 : weather.ordinal();
     }
