+++ runtime/src/main/java/me/shedaniel/rei/impl/common/entry/type/types/BuiltinEntryDefinition.java
@@ -95,7 +95,7 @@ public class BuiltinEntryDefinition<T> implements EntryDefinition<T>, EntrySeria
     }
     
     @Override
-    public int hash(EntryStack<T> entry, T value, ComparisonContext context) {
+    public long hash(EntryStack<T> entry, T value, ComparisonContext context) {
         return empty ? 0 : Objects.hash(value.getClass().getName(), value);
     }
